@@ -12,7 +12,7 @@ return [
      */
     'disk' => 'public',
 
-    /**
+    /**w
      * User agent rules for different paths
      */
     'user_agents' => [
